Allow service name per span as an option (#53)
* Allow service name per span as an option

In some scenarios it is desirable to allow service name to be set per span and not globally by process. This change adds a env var to allow that option.

* Ensure instrumentations follow the service name pattern

Before these "custom integration service names" were hidden because the span service name was ignored. Now we want these to follow the same as the process, otherwise, service name for these instrumentation will change if combined with manual instrumentation and the option to set service name per span.

* Add _ENABLED suffix to env var and info to README