Enable WCF integration (#54)
Enabling WCF instrumentation that already exists on DD. This is to just enable what already exists. There were no automated tests for this integration, this PR does not add one for it. All validations were made manually by enabling profiling for Samples.Wcf and Samples.WcfClient (they are built when running the cmd to test integrations).

The client-side for WSHttpBinding and BasicHttpBinding generates client spans via the WebRequest instrumentation. Enabling WCF integration creates a span for the server-side. The context is properly flowing between the client and the server. There is no client-side span for NetTcpBinding.

The server span looks like a generic HTTP server-side span in order to make clear that this is a WCF span the operation name will be pre-pended with wcf.request.