Managed loader cannot resolve profiler managed libraries when installed by the MSI installer (#1820)
### Context

When installing the profiler and the tracer using the MSI installer, the profiler managed library cannot be resolved.

To resolve the profiler managed library, the managed loader uses the `COR_PROFILER_PATH_XX` or `CORECLR_PROFILER_PATH_XX` to locate the profiler home directory.

But, the MSI installer does not populate those environment variables, instead it registers the CLR profiler CLSID and the path in the registry.

### Fixes

Changes proposed in this pull request:

To resolve the tracer managed library, the managed loader relies on the `DD_DOTNET_TRACER_HOME` environment variable. (even if today the tracer does not use the managed loader).

We can introduction a `DD_DOTNET_PROFILER_HOME` environment variable to replicate the same pattern used for the tracer.