Add a separate performance pipeline for running Performance tests (#1842)
* Add a separate performance pipeline for running Benchmarks and Throughput tests

The advantage of this is that we can make the benchmark and throughput results optional. We still want to run them for every PR, but it means in cases where a rebase or a trivial change is required, we wouldn't need to wait for the performance tests to finish before merging.

* Update benchmark comparison tool to use correct pipeline ID

* Need to have a stage without any dependencies

* depends_on isn't required, as shuoldn't be triggered until source completes

* Make stages run in parallel