Create a shared `CurrentProcess` helper class in based on the Tracer's internal helper. (#1876)
The Tracer has an (internal) `ProcessHelpers` utility class. We recently copied some code out of it for use in the Profiler.
We already know that CodeHotspots will also need functionality of the Process API.

This change sets up up a helper class in the shared code folder that covers the same functionality. To avoid cross-PR dependencies, it is not yet used, but we will start using it in the Profiler after this PR is merged. The Tracer team should consider whether or not they like using the shared code too, and decide based on their preferences.