Run log injection tests on linux (#1929)
* Enable running Logs Injection tests on Linux

We really need to be running these so that we are testing multiple package versions

* Pass package version in LogInjection integration tests

We need to pass the package version so that we can get the path to the correct version of the application, when building the multi-package samples. Otherwise we are checking the wrong log files!

* Try and workaround NLog on Linux issues

No matter what I try, no matter which directory char we use etc, I just can't get NLog to behave when reading the xml config. As a workaround, just force it to put the logs in the correct place on linux by changing it in code

* Use older "default" versions for logs injection on .NET Framework

Windows only currently runs a single package version for integration tests. To increase the number of covered versions for logs-injection, use on of the .NET Framework-only packages when running on full framework.

* Fix NLog for configuration for NLog 4.6+

In NLog4.6+, LibLog adds the attributes to the MDLC instead of the MDC, so we need to use a new config.