Small build fixes (#1936)
* Fix build_in_docker.sh

This was broken when we moved things into the tracer folder and wasn't fixed correctly

* Always build AspNetCoreRazorPages when building single sample

If you try and build a single sample on linux using (for example)

```
./tracer/build_in_docker.sh Clean BuildTracerHome BuildAndRunLinuxIntegrationTests -SampleName "LogsInjection.NLog"
```

then the build will fail. That's because the linux integration tests depend on `Samples.AspNetCoreRazorPages`, but when a single sample name is provided, it won't be built.

This change ensures we _always_ build `Samples.AspNetCoreRazorPages`, even when a specific sample is specified using `-SampleName <name>`