Refactor reverse duck-typing and add negative tests (#1900)
* Add tests that we _can't_ duck type invalidly defined types

We have lots of tests that we _can_ duck type certain things, but there are many cases where we should throw when attempting to Duck Cast. If we don't, we risk accidentally thinking we've duck cast correctly when we haven't.

Obviously an even better approach would be to have an analyzer that detects this too, but that's a _lot_ more work

* Add checks for method resolution that Duck attribute works with generic parameter names

* Add some error tests for reverse duck typing

* Add failing test for reverse proxy where virtual methods already have a method body

* Add test that reverse-proxy properties work correctly (they don't currently)

* Fix uninitialized fields and properties in reverse proxy

Call base constructor if the proxy type is inheriting from a class or abstract class.

* Change type returned by DuckType.Create() to an object

If the ProxyType is a struct, then even if the implementation is correct, this will throw for struct proxies when it attempts to cast to an IDuckType (as the struct doesn't implement IDuckType). The cast is largely unnecessary anyway, and can be cast explicitly by the caller if necessary (and safe)

* Fix cases in field and property duck typing that were being too lenient

They were being lenient to handle _reverse_ duck typing scenarios, but we'll split those out for clarity and increase the strictness here

Also add some warnings where we don't (can't) know if a conversion is valid, so it would result in a runtime error accessing the member

* Add additional guard clauses for method duck typing

- Catch when an argument changes from generic to non-generic in a generic method
- Catch when the original method returns void, but the new method returns non-void

* Skip tests for which we can't (or won't) detect failures

(Primarily around return types and conversions)

* Split reverse-duck-typing from forward-duck-typing

While these cases are quite similar, there are a bunch of requirements and edge cases that make interleaving the implementation problematic. Some of the additional tests added highlighted this problem.

To try and make changes to the duck-typing code safer, this separates both the implementation and the API of reverse duck typing from "forward" duck typing. This also fixes some of the implementation issues that were previously causing failures

* Make [DuckReverseMethod] a mirror of [Duck] attribute

- Extracted the common features to a DuckAttributeBase
- Allow [DuckReverseMethod] to be used on properties
- Update method resolution to use the new attributes
- Fix method resolution for forward duck-typing when using Duck attribute and have generic parameters
- Fix method resolution for reverse duck-typing
- [DuckIgnore] is no longer used by reverse duck typing, as you must decorate all your reverse methods/properties anyway

* Fix [DuckReverseMethod] usages

* Update SelectTargetMethod implementation

- Fix case where the duck type attribute parameters don't match method parameters count
- Fix duck chaining from reverse duck types - need to do _reverse_ duck chaining here, not normal forward duck chaining
- Fix reverse duck typing not finding
- Fix invalid test (caught with above changes)

* Handle additional incorrect usage case and add additional exception tests

* Ensure IntegrationOptions catches DuckType exceptions when using DuckCast(Type), as well as DuckCast<T>

* Refactor DuckType.Methods.cs to reduce duplication

* Update Public API tests

This added some public DuckTyping methods. They could be made internal, but their equivalents are already public, so I think it makes more sense to keep them all public for now

* Update ExceptionsTests

Introduce ReverseProxyMustImplementGenericMethodAsGenericException to make it easier to debug the issue
Remove unused exception
Add missing tests for ReverseProxyMissingMethodImplementationException
Add some extra tests for code coverage purposes

* Update incorrect comments

* Fix public API

* Comments from code review

* Update reverse proxy properties implementation

Reverse proxies for methods was previously broken.

As for methods, when we're doing reverse proxying, we need to do the opposite ducktype/duckcast compared to forward proxying

This also removes the DuckKind.Field for reverse proxies (as that doesn't make sense)
Also, updated the "duck cast/duck chain" delegates to return the actual type that you end up with, for use by the caller.

* Handle case where [DuckReverseMethod] attribute has different number of parameters than method

This always points to a configuration error for reverse duck typing, whereas for forward duck typing this isn't _necessarily_ an error (as the target may have optional parameters, for example)

Add an explicit check for the reverse duck type case to catch this.

* Fix incorrect arguments in test

* Use public types so we can test net452 in reverse proxy error tests

* Fix incorrectly ignored `WrongArgumentTypeInterfaceImplementations.DuckChainReturnMethod` test

* Fix Public API tests

* Ensure we catch _all_ exceptions when creating a duck type

Co-authored-by: Tony Redondo <tony.redondo@datadoghq.com>