Remove all LibLog code and simplify ScopeManager code (#2184)
In Tracer 2.0, all automatic logs injection is done via automatic instrumentation. This means we no longer have a need for the LibLog library and the additional complexity it introduced in the Tracer code. This removes all of that, including the ScopeManager event handlers that the LibLogScopeEventSubscriber relied on to perform correct logs correlation.