Fix no-op pipeline running when it shouldn't (#2334)
* Remove wildcards from path triggers

These are not supported, and will just be ignored

* Update no-op pipeline to only run when it should

Azure Devops doesn't support the sort of include/exclude logic we need, so we have to manually calculate it ourselves. This is obviously far from ideal, but currently we're running the Noop pipeline when _any_ file in the consolidated-pipeline exclude is touched, not ONLY when they're touched.

* Delete unused file