Add internal "start" api to `Tracer`/`TracerManager` (#2330)
* Add a mechanism to not start tracermanager processes until after Tracer.Instance is set

This is used to avoid issues related to instrumenting types that we use as part of our infrastructure.

For example, this requirement was identified as part of the Telemetry work, in which calling Tracer.Instance initializes a TracerManager, and the TelemetryController immediately starts it's "push" loop. This loop uses HttpClient, which is auto-instrumented, and hence calls `Tracer.Instance` as part of the instrumentation (before checking for "do not trace" headers).

This call to Tracer.Instance all happens as part of the _same_ call stack that called the _initial_ Tracer.Instance, and hence is reentrant (and the `lock` does nothing to stop it). The end result is that we have _two_ TracerManager instances, which is a situation we were trying to avoid.

Another (additional) solution would be to configure a `DoNotTrace` class, but that has more general applicability, probably shouldn't be part of Datadog.Trace.dll, and needs to be publicly exposed, so this is somewhat of a stop gap.

* Don't start the direct log submission sink until after Tracer.Instance set

To avoid potential initialization issues, require calling `Start()` on BatchingSink before it starts the main emit loop.