Update GraphQL and Security ASP.NET Core tests to reduce flakiness (#2451)
* Update GraphQLTests to reduce flakiness

Previously, we were trying to find an open port, and then use that for all requests.

With this change, we rely on ASP.NET Core's own "find an open port" mechanism, and then read that port from the output. This _should_ be more reliable, and hopefully will reduce the number of cases of "unable to bind port" that we see

* Update Security ASP.NET Core tests to reduce flakiness

Previously, we were trying to find an open port, and then use that for all requests.

With this change, we rely on ASP.NET Core's own "find an open port" mechanism, and then read that port from the output. This _should_ be more reliable, and hopefully will reduce the number of cases of "unable to bind port" that we see

* Update tracer/test/Datadog.Trace.ClrProfiler.IntegrationTests/GraphQLTests.cs

Co-authored-by: Kevin Gosse <krix33@gmail.com>

Co-authored-by: Kevin Gosse <krix33@gmail.com>