Refactor no-op pipeline to not use so many hosts (#2467)
Currently the no-op pipeline is generating a different job for each github status update check it needs to run. As we now have the profiler in there as well, the no-op pipeline is running on basically every PR. That means 20ish agents are required to run these checks. While they only run for a very short period of time (~2s), there's still the provisioning time etc to think about.

With this approach, instead of using a matrix to generate multiple jobs, we output a single semi-colon delimited list of jobs to update in GitHub, and do it all in one hit.

This also added some refactoring of the GenerateNoopStages code. Functionally it's basically the same, but we output a list of jobs instead of a json blob