[Profiler] Implement new pipeline to generate .pprof via libddprof (#2524)
Build a new native only pipeline allowing multiple sample collectors (only WallTime in the PR) to provide samples to the aggregator that generates and send .pprof file via libddprof.

Note: the integration is connected to a feature flag set in the DD_INTERNAL_PROFILING_LIBDDPROF_ENABLED environment variable

* Make native .pprof generation the default (#2515)
* fix Visual Studio launchSettings.json to run in the tracer repository
* Fix unit test for libddprof usage during .pprof generation
* Make the old profile exporter pipeline the default one

We previously set the new profile exporter as the default one.
But this has not been sufficiently tested.
Make the old profile exporter pipeline the default and add tests for the new one:
- Address sanitizer
- Smoke tests
- Benchmark tests

Co-authored-by: chrisnas <chrisnas@users.noreply.github.com>
Co-authored-by: Gregory LEOCADIE <gregory.leocadie@datadoghq.com>