Upstream sync part3 (#443)
* Upstream sync 4d5418990c88d209946369c5b49e4a945a00ed1c..6f1635f2181159e96c2a83c1c217440b08ba8add

* Add better handling of nullable type chaining (#2424)

* Add better handling of nullable type chaining

Currently, when doing duck chaining, we call `CreateCache<T>.Create(object instance)`
 to do the conversion. Note that this is a boxing operation.
Inside this method, we then call `instance.GetType()` to determine
the type to convert from. However, [doing a boxing operation on a non-null nullable type converts it to the type itself](https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/builtin-types/nullable-value-types#how-to-identify-a-nullable-value-type).
The end result is that we create the incorrect proxy mapping.

## An example of the problem

For example, consider this type hierarchy:

```csharp
public class GenericTestClass
{
    public TestStruct? MyTestStruct { get; set; }
}

public struct TestStruct { }
```

Which we want to create a proxy for:

```csharp
[DuckCopy]
public struct GenericProxyStruct
{
    public NullableProxyTestStruct MyTestStruct;
}

[DuckCopy]
public struct NullableProxyTestStruct
{
    public ProxyTestStruct Value;
}

[DuckCopy]
public struct ProxyTestStruct { }
```

Running the following before this PR would throw, as it trys
to proxy `TestStruct` to `NullableProxyTestStruct` _not_ `TestStruct?`

```csharp
var instance = new GenericTestClass { MyTestStruct = new TestStruct() };
var proxy = instance.DuckCast<GenericProxyStruct>();
```

> Note that we currently require an "intermediate" proxy to handle the nullable. It would be nice to handle this implicitly in the future, which incidentally would also solve this issue

The biggest problem is that the behaviour is _different_ depending
on whether the proxied property is `null` at runtime, or it has a value,
so there's no "safe" way to create a proxy heirarchy that works in all cases.

## The fix

The source of this problem is that the `CreateCache<T>.Create(object instance)`
call that is part of `DuckChaining` boxes the source field.
To work around the issue, I created a new, generic, method
`CreateCache<T>.CreateFrom<T>(T instance)` which avoids the boxing. This then creates a proxy between the nullable type as expected,
and the above hierarchy works as expected.

This fix only helps with the duck-chaining case, but this is the
main case in which errors can occur - in most cases, _directly_
duck typing a nullable isn't something we'll need to do (hopefully!)

* No need to box if duck typing a value type with CreateFrom

* If the `int` status code is not valid, don't set it (#2529)

* If the integer status code is not valid, don't set it

It will get stripped for being invalid during tag normalization anyway

* Skip large values too

* Managed loader cache (#2533)

* Add a cache layer to avoid File.Exists and loading the assembly multiple times.

* remove log entry and array initialization

* [Profiler] Add profiler in the automatic version bump process (#2530)

* Add version.h file to the project file
* Update version in CMakeLists.txt
+ set BetaVersion to 1 (we are currently at 1)

* [Profiler] Implement new pipeline to generate .pprof via libddprof (#2524)

Build a new native only pipeline allowing multiple sample collectors (only WallTime in the PR) to provide samples to the aggregator that generates and send .pprof file via libddprof.

Note: the integration is connected to a feature flag set in the DD_INTERNAL_PROFILING_LIBDDPROF_ENABLED environment variable

* Make native .pprof generation the default (#2515)
* fix Visual Studio launchSettings.json to run in the tracer repository
* Fix unit test for libddprof usage during .pprof generation
* Make the old profile exporter pipeline the default one

We previously set the new profile exporter as the default one.
But this has not been sufficiently tested.
Make the old profile exporter pipeline the default and add tests for the new one:
- Address sanitizer
- Smoke tests
- Benchmark tests

Co-authored-by: chrisnas <chrisnas@users.noreply.github.com>
Co-authored-by: Gregory LEOCADIE <gregory.leocadie@datadoghq.com>

* Use managed code to generate pprof file (#2540)

By default we still want to use the Managed code to generate pprof file.

* DD_INTERNAL_PROFILING_LIBDDPROF_ENABLED -> SIGNALFX_INTERNAL_PROFILING_LIBDDPROF_ENABLED

* stop propagating `x-datadog-tags` header (#2507)

* [CIApp] - Add support for XUnit LogsDirectSubmission (#2531)

* - Add support for XUnit Integration in DirectLogSubmission
- Add Flush support in DirectLogSubmission
- Add CI Visibility configuration settings.

* Add TestOutputHelper.QueueTestOutput integration.

* change comments.

* Remove other loggers than XUnit.

* Update tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/Testing/XUnit/XUnitTestOutputHelperQueueTestOutputIntegration.cs

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>

* Update tracer/src/Datadog.Trace/ClrProfiler/AutoInstrumentation/Testing/XUnit/XUnitTestOutputHelperQueueTestOutputIntegration.cs

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>

* Checks and changes.

* Improve flush.

* Fix build issues.

* Add _dd.origin tag to the logs.

* Add Tests.

* fix test suite

* Fix tests and mock

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>

* [Profiler] Fix build_windows_profiler AzOp step (#2541)

* Restore vcxproj first

* Allow disabling samples on Alpine, and always running only major versions (#2534)

* Add additional package version testing options

- Allow always using major package versions, even when minor is requested. This is necessary when there are a _lot_ of minor versions (e.g. GRPC) which would cause the number of samples being built to balloon.
- Allow using a different set of versions when running on Alpine. Specifically, currently allow _skipping_ running on Alpine.

* Run GeneratePackageVersions

* Fix CreatePlatformlessSymlinks target

* Rework alpine package handling

Co-authored-by: Andrew Lock <andrew.lock@datadoghq.com>
Co-authored-by: Tony Redondo <tony.redondo@datadoghq.com>
Co-authored-by: Gregory LEOCADIE <gregory.leocadie@datadoghq.com>
Co-authored-by: chrisnas <chrisnas@users.noreply.github.com>
Co-authored-by: Zach Montoya <zach.montoya@datadoghq.com>