Add support for UDS to telemetry transport (#2617)
* Add XML docs around using http://localhost for base endpoint with UDS/Named pipes

* Add explicit configuration key for enabling/disabling agentless telemetry

* Refactor TelemetrySettings so they are always in a consistent state

i.e. discriminated union of agentless settings or agent settings

* Update Telemetry transport to reuse IApiRequestFactory

Includes adding support for named pipes and UDS

* Build in support for telemetry testing into `MockTracerAgent`

When we're sending to the "agent" endpoint, we need to reuse the same Socket/HttpListener for both traces and telemetry, so we need to incorporate the `MockTelemetryAgent` functionality into the `MockTracerAgent`, and split based on path.

This required adding support for parsing the URL requested from the HTTP request in `MockHttpParser`.

* Fix TraceAnnotationsVersionMismatchNewerNuGetTests

This test currently uses the _old_ env vars for configuring telemetry, so need to use those
We can't use the MockAgent-based telemetry because the older nuget doesn't support agent-based telemetry, only agentless

* PR Feedback

* Set Connection: Close on responses to try and fix flakiness in UDS

minor: switch to StringBuilder