Exclude guid-like assembly names from telemetry (#2954)
We're seeing a lot of dependencies listed that are just GUIDs. These aren't useful and blow up cardinality, so exclude them from the dependency list.

Went for a _super_ basic approach that just checks it's a string of the right length, with four hyphens in the right place. We could be more rigorous if people prefer, but these seems good enough to me