Add FNV-64bit hash implementation for Data Streams Monitoring (#2973)
* Add simple FNV hash implementation

Required for upcoming data stream monitoring work.
There are separate implementations for string, byte, and Span<byte> for the various different usages we will require.

* Make hash generation with strings allocation free for small strings

* Use SkipLocalsInit to speed up the stackalloc hash generation

* Add support for incrementally generating a hash

This allows you to pass the "previous" hash version to the GenerateHash version. This will be useful as we need to concatenate multiple strings together to create the final hash value. By using these "incremental" versions, we can generate the hash in a zero-allocation way (for small enough strings in netcoreapp3.1+).